Fixes as requested by Fabian in Review.
