+++ tool_packages/unique_internal_search/tests/test_service.py
@@ -1547,3 +1547,250 @@ class TestInternalSearchTool:
             # Since we can't easily check the exact state value due to ProgressState enum,
             # verify that multiple calls were made (at least one should be FINISHED)
             assert mock_tool_progress_reporter.notify_from_tool_call.call_count >= 2
+
+    @pytest.mark.ai
+    def test_define_reference_list__returns_list__with_internal_chunks(
+        self,
+        sample_content_chunk: ContentChunk,
+    ) -> None:
+        """
+        Purpose: Verify define_reference_list returns list of references for internal search.
+        Why this matters: References link log entries to internal document sources.
+        Setup summary: Create internal content chunk, call define_reference_list, verify returns list.
+        """
+        from unique_toolkit.content.schemas import ContentReference
+
+        from unique_internal_search.service import InternalSearchService
+
+        # Arrange
+        content_chunks = [sample_content_chunk]
+        data: list[ContentReference] = []
+
+        # Act
+        references = InternalSearchService.define_reference_list(
+            source="internal", content_chunks=content_chunks, data=data
+        )
+
+        # Assert
+        assert isinstance(references, list)
+        assert len(references) == 1
+
+    @pytest.mark.ai
+    def test_define_reference_list__sets_sequence_number__for_first_chunk(
+        self,
+        sample_content_chunk: ContentChunk,
+    ) -> None:
+        """
+        Purpose: Verify define_reference_list sets sequence_number to 0 for first chunk.
+        Why this matters: Sequence numbers order references in display.
+        Setup summary: Create internal content chunk, call define_reference_list, verify sequence_number is 0.
+        """
+        from unique_toolkit.content.schemas import ContentReference
+
+        from unique_internal_search.service import InternalSearchService
+
+        # Arrange
+        content_chunks = [sample_content_chunk]
+        data: list[ContentReference] = []
+
+        # Act
+        references = InternalSearchService.define_reference_list(
+            source="internal", content_chunks=content_chunks, data=data
+        )
+
+        # Assert
+        assert isinstance(references[0].sequence_number, int)
+        assert references[0].sequence_number == 0
+
+    @pytest.mark.ai
+    def test_define_reference_list__sets_source_id__from_chunk_id(
+        self,
+        sample_content_chunk: ContentChunk,
+    ) -> None:
+        """
+        Purpose: Verify define_reference_list sets source_id from chunk ID.
+        Why this matters: Source ID identifies the internal document chunk.
+        Setup summary: Create internal content chunk with ID, call define_reference_list, verify source_id matches.
+        """
+        from unique_toolkit.content.schemas import ContentReference
+
+        from unique_internal_search.service import InternalSearchService
+
+        # Arrange
+        content_chunks = [sample_content_chunk]
+        data: list[ContentReference] = []
+
+        # Act
+        references = InternalSearchService.define_reference_list(
+            source="internal", content_chunks=content_chunks, data=data
+        )
+
+        # Assert
+        assert isinstance(references[0].source_id, str)
+        assert references[0].source_id == "chunk_123"
+
+    @pytest.mark.ai
+    def test_define_reference_list__sets_name__from_chunk_key_when_no_title(
+        self,
+    ) -> None:
+        """
+        Purpose: Verify define_reference_list uses key when title is missing.
+        Why this matters: References need names even when title is unavailable.
+        Setup summary: Create internal content chunk without title, call define_reference_list, verify name is key.
+        """
+        from unique_toolkit.content.schemas import ContentChunk, ContentReference
+
+        from unique_internal_search.service import InternalSearchService
+
+        # Arrange
+        content_chunk = ContentChunk(
+            id="chunk_internal_2",
+            text="Another internal document",
+            key="doc_key_2",
+        )
+        content_chunks = [content_chunk]
+        data: list[ContentReference] = []
+
+        # Act
+        references = InternalSearchService.define_reference_list(
+            source="internal", content_chunks=content_chunks, data=data
+        )
+
+        # Assert
+        assert isinstance(references[0].name, str)
+        assert references[0].name == "doc_key_2"
+
+    @pytest.mark.ai
+    def test_define_reference_list__sets_name__from_chunk_title_when_available(
+        self,
+    ) -> None:
+        """
+        Purpose: Verify define_reference_list uses title when available.
+        Why this matters: Title provides meaningful name for internal document references.
+        Setup summary: Create internal content chunk with title, call define_reference_list, verify name is title.
+        """
+        from unique_toolkit.content.schemas import ContentChunk, ContentReference
+
+        from unique_internal_search.service import InternalSearchService
+
+        # Arrange
+        content_chunk = ContentChunk(
+            id="chunk_internal_1",
+            text="Internal document content",
+            title="Internal Document Title",
+            key="doc_key_1",
+        )
+        content_chunks = [content_chunk]
+        data: list[ContentReference] = []
+
+        # Act
+        references = InternalSearchService.define_reference_list(
+            source="internal", content_chunks=content_chunks, data=data
+        )
+
+        # Assert
+        assert isinstance(references[0].name, str)
+        assert references[0].name == "Internal Document Title"
+
+    @pytest.mark.ai
+    def test_define_reference_list__sets_empty_url__for_internal_chunks(
+        self,
+        sample_content_chunk: ContentChunk,
+    ) -> None:
+        """
+        Purpose: Verify define_reference_list sets empty URL for internal chunks.
+        Why this matters: Internal documents don't have URLs, only source IDs.
+        Setup summary: Create internal content chunk, call define_reference_list, verify URL is empty string.
+        """
+        from unique_toolkit.content.schemas import ContentReference
+
+        from unique_internal_search.service import InternalSearchService
+
+        # Arrange
+        content_chunks = [sample_content_chunk]
+        data: list[ContentReference] = []
+
+        # Act
+        references = InternalSearchService.define_reference_list(
+            source="internal", content_chunks=content_chunks, data=data
+        )
+
+        # Assert
+        assert isinstance(references[0].url, str)
+        assert references[0].url == ""
+
+    @pytest.mark.ai
+    def test_define_reference_list__increments_sequence_number__for_multiple_chunks(
+        self,
+    ) -> None:
+        """
+        Purpose: Verify define_reference_list increments sequence_number for multiple chunks.
+        Why this matters: Multiple references must be ordered correctly.
+        Setup summary: Create two internal content chunks, call define_reference_list, verify sequence numbers increment.
+        """
+        from unique_toolkit.content.schemas import ContentChunk, ContentReference
+
+        from unique_internal_search.service import InternalSearchService
+
+        # Arrange
+        content_chunks = [
+            ContentChunk(
+                id="chunk_1",
+                text="First chunk",
+                key="doc1.pdf",
+            ),
+            ContentChunk(
+                id="chunk_2",
+                text="Second chunk",
+                key="doc2.pdf",
+            ),
+        ]
+        data: list[ContentReference] = []
+
+        # Act
+        references = InternalSearchService.define_reference_list(
+            source="internal", content_chunks=content_chunks, data=data
+        )
+
+        # Assert
+        assert isinstance(references[0].sequence_number, int)
+        assert references[0].sequence_number == 0
+        assert isinstance(references[1].sequence_number, int)
+        assert references[1].sequence_number == 1
+
+    @pytest.mark.ai
+    def test_define_reference_list__skips_chunks__with_empty_name(
+        self,
+    ) -> None:
+        """
+        Purpose: Verify define_reference_list skips chunks with empty name (no title and no key).
+        Why this matters: Only chunks with valid names should be included in references.
+        Setup summary: Create chunks with empty name, call define_reference_list, verify they are skipped.
+        """
+        from unique_toolkit.content.schemas import ContentChunk, ContentReference
+
+        from unique_internal_search.service import InternalSearchService
+
+        # Arrange
+        content_chunks = [
+            ContentChunk(
+                id="chunk_1",
+                text="Content with key",
+                key="doc1.pdf",
+            ),
+            ContentChunk(
+                id="chunk_2",
+                text="Content without name",
+                key="",
+            ),
+        ]
+        data: list[ContentReference] = []
+
+        # Act
+        references = InternalSearchService.define_reference_list(
+            source="internal", content_chunks=content_chunks, data=data
+        )
+
+        # Assert
+        assert len(references) == 1
+        assert references[0].name == "doc1.pdf"
