+++ unique_toolkit/tests/agentic/test_logger_service.py
@@ -28,7 +28,7 @@ from unique_toolkit.chat.schemas import (
     MessageLogUncitedReferences,
 )
 from unique_toolkit.chat.service import ChatService
-from unique_toolkit.content.schemas import ContentChunk, ContentReference
+from unique_toolkit.content.schemas import ContentReference
 
 
 # Centralized Fixtures
@@ -103,50 +103,6 @@ def reset_message_order_counters():
 
     # Clear counters after test
     _request_counters.clear()
-
-
-@pytest.fixture
-def web_content_chunk() -> ContentChunk:
-    """
-    Purpose: Provide a web search content chunk for reference list tests.
-    Why this matters: Ensures consistent test data for web search scenarios.
-    Setup summary: Creates ContentChunk with URL and title for web search.
-    """
-    return ContentChunk(
-        id="chunk_1",
-        text="Content from web page 1",
-        url="https://example.com/page1",
-        title="Page 1 Title",
-    )
-
-
-@pytest.fixture
-def internal_content_chunk_with_title() -> ContentChunk:
-    """
-    Purpose: Provide an internal content chunk with title for reference tests.
-    Why this matters: Tests behavior when title is available for internal references.
-    Setup summary: Creates ContentChunk with title and key for internal search.
-    """
-    return ContentChunk(
-        id="chunk_internal_1",
-        text="Internal document content",
-        title="Internal Document Title",
-        key="doc_key_1",
-    )
-
-
-@pytest.fixture
-def internal_content_chunk_without_title() -> ContentChunk:
-    """
-    Purpose: Provide an internal content chunk without title for reference tests.
-    Why this matters: Tests fallback behavior when title is missing.
-    Setup summary: Creates ContentChunk with only key, no title.
-    """
-    return ContentChunk(
-        id="chunk_internal_2",
-        text="Another internal document",
-        key="doc_key_2",
-    )
 
 
 # Initialization Tests
@@ -235,7 +191,9 @@ def test_get_next_message_order__starts_from_one__with_different_message_id_AI()
     """
     # Arrange
     first_message_id = "test_message_one"
-    _ = MessageStepLogger.get_next_message_order(message_id=first_message_id)  # Increment first
+    _ = MessageStepLogger.get_next_message_order(
+        message_id=first_message_id
+    )  # Increment first
     different_message_id = "test_message_two"
 
     # Act
@@ -492,300 +450,6 @@ def test_write_message_log_text_message__increments_order__on_multiple_calls_AI(
     assert calls[1].kwargs["order"] == 2  # type: ignore[attr-defined]
     assert isinstance(calls[2].kwargs["order"], int)  # type: ignore[attr-defined]
     assert calls[2].kwargs["order"] == 3  # type: ignore[attr-defined]
-
-
-# Reference List Tests - Web Search
-@pytest.mark.ai
-def test_define_reference_list__returns_list__with_web_search_chunks_AI(
-    web_content_chunk: ContentChunk,
-) -> None:
-    """
-    Purpose: Verify define_reference_list returns list of references for web search.
-    Why this matters: References link log entries to web search result URLs.
-    Setup summary: Create web content chunk, call define_reference_list, verify returns list.
-    """
-    # Arrange
-    content_chunks = [web_content_chunk]
-    data = []
-
-    # Act
-    references = MessageStepLogger.define_reference_list(
-        source="web_search", content_chunks=content_chunks, data=data
-    )
-
-    # Assert
-    assert isinstance(references, list)
-    assert len(references) == 1
-
-
-@pytest.mark.ai
-def test_define_reference_list__sets_sequence_number__for_first_chunk_AI(
-    web_content_chunk: ContentChunk,
-) -> None:
-    """
-    Purpose: Verify define_reference_list sets sequence_number to 1 for first chunk.
-    Why this matters: Sequence numbers order references in display.
-    Setup summary: Create web content chunk, call define_reference_list, verify first reference sequence_number is 1.
-    """
-    # Arrange
-    content_chunks = [web_content_chunk]
-    data = []
-
-    # Act
-    references = MessageStepLogger.define_reference_list(
-        source="web_search", content_chunks=content_chunks, data=data
-    )
-
-    # Assert
-    assert isinstance(references[0].sequence_number, int)
-    assert references[0].sequence_number == 0
-
-
-@pytest.mark.ai
-def test_define_reference_list__sets_url__from_chunk_url_AI(
-    web_content_chunk: ContentChunk,
-) -> None:
-    """
-    Purpose: Verify define_reference_list sets URL from chunk URL.
-    Why this matters: URL links reference to source web page.
-    Setup summary: Create web content chunk with URL, call define_reference_list, verify reference URL matches.
-    """
-    # Arrange
-    content_chunks = [web_content_chunk]
-    data = []
-
-    # Act
-    references = MessageStepLogger.define_reference_list(
-        source="web_search", content_chunks=content_chunks, data=data
-    )
-
-    # Assert
-    assert isinstance(references[0].url, str)
-    assert references[0].url == "https://example.com/page1"
-
-
-@pytest.mark.ai
-def test_define_reference_list__sets_source__from_parameter_AI(
-    web_content_chunk: ContentChunk,
-) -> None:
-    """
-    Purpose: Verify define_reference_list sets source from parameter.
-    Why this matters: Source identifies where reference content originated.
-    Setup summary: Create web content chunk, call define_reference_list with source, verify reference source matches.
-    """
-    # Arrange
-    content_chunks = [web_content_chunk]
-    source = "web_search"
-    data = []
-
-    # Act
-    references = MessageStepLogger.define_reference_list(
-        source=source, content_chunks=content_chunks, data=data
-    )
-
-    # Assert
-    assert isinstance(references[0].source, str)
-    assert references[0].source == source
-
-
-@pytest.mark.ai
-def test_define_reference_list__sets_name__from_chunk_url_AI(
-    web_content_chunk: ContentChunk,
-) -> None:
-    """
-    Purpose: Verify define_reference_list sets name from chunk URL.
-    Why this matters: Name is displayed to users as reference identifier.
-    Setup summary: Create web content chunk with URL, call define_reference_list, verify reference name is URL.
-    """
-    # Arrange
-    content_chunks = [web_content_chunk]
-    data = []
-
-    # Act
-    references = MessageStepLogger.define_reference_list(
-        source="web_search", content_chunks=content_chunks, data=data
-    )
-
-    # Assert
-    assert isinstance(references[0].name, str)
-    assert references[0].name == "https://example.com/page1"
-
-
-@pytest.mark.ai
-def test_define_reference_list__increments_sequence_number__for_multiple_chunks_AI() -> (
-    None
-):
-    """
-    Purpose: Verify define_reference_list increments sequence_number for multiple chunks.
-    Why this matters: Multiple references must be ordered correctly.
-    Setup summary: Create two web content chunks, call define_reference_list, verify sequence numbers increment.
-    """
-    # Arrange
-    content_chunks = [
-        ContentChunk(
-            id="chunk_1",
-            text="Content from web page 1",
-            url="https://example.com/page1",
-            title="Page 1 Title",
-        ),
-        ContentChunk(
-            id="chunk_2",
-            text="Content from web page 2",
-            url="https://example.com/page2",
-            title="Page 2 Title",
-        ),
-    ]
-    data = []
-
-    # Act
-    references = MessageStepLogger.define_reference_list(
-        source="web_search", content_chunks=content_chunks, data=data
-    )
-
-    # Assert
-    assert isinstance(references[0].sequence_number, int)
-    assert references[0].sequence_number == 0
-    assert isinstance(references[1].sequence_number, int)
-    assert references[1].sequence_number == 1
-
-
-# Reference List Tests - Internal Search
-@pytest.mark.ai
-def test_define_reference_list_for_internal__returns_list__with_internal_chunks_AI(
-    internal_content_chunk_with_title: ContentChunk,
-) -> None:
-    """
-    Purpose: Verify define_reference_list_for_internal returns list of references.
-    Why this matters: References link log entries to internal document sources.
-    Setup summary: Create internal content chunk, call define_reference_list_for_internal, verify returns list.
-    """
-    # Arrange
-    content_chunks = [internal_content_chunk_with_title]
-    data = []
-
-    # Act
-    references = MessageStepLogger.define_reference_list_for_internal(
-        source="internal", content_chunks=content_chunks, data=data
-    )
-
-    # Assert
-    assert isinstance(references, list)
-    assert len(references) == 1
-
-
-@pytest.mark.ai
-def test_define_reference_list_for_internal__sets_sequence_number__for_first_chunk_AI(
-    internal_content_chunk_with_title: ContentChunk,
-) -> None:
-    """
-    Purpose: Verify define_reference_list_for_internal sets sequence_number to 0.
-    Why this matters: Sequence numbers order references in display.
-    Setup summary: Create internal content chunk, call define_reference_list_for_internal, verify sequence_number is 0.
-    """
-    # Arrange
-    content_chunks = [internal_content_chunk_with_title]
-    data = []
-
-    # Act
-    references = MessageStepLogger.define_reference_list_for_internal(
-        source="internal", content_chunks=content_chunks, data=data
-    )
-
-    # Assert
-    assert isinstance(references[0].sequence_number, int)
-    assert references[0].sequence_number == 0
-
-
-@pytest.mark.ai
-def test_define_reference_list_for_internal__sets_source_id__from_chunk_id_AI(
-    internal_content_chunk_with_title: ContentChunk,
-) -> None:
-    """
-    Purpose: Verify define_reference_list_for_internal sets source_id from chunk ID.
-    Why this matters: Source ID identifies the internal document chunk.
-    Setup summary: Create internal content chunk with ID, call define_reference_list_for_internal, verify source_id matches.
-    """
-    # Arrange
-    content_chunks = [internal_content_chunk_with_title]
-    data = []
-
-    # Act
-    references = MessageStepLogger.define_reference_list_for_internal(
-        source="internal", content_chunks=content_chunks, data=data
-    )
-
-    # Assert
-    assert isinstance(references[0].source_id, str)
-    assert references[0].source_id == "chunk_internal_1"
-
-
-@pytest.mark.ai
-def test_define_reference_list_for_internal__sets_name__from_chunk_title_AI(
-    internal_content_chunk_with_title: ContentChunk,
-) -> None:
-    """
-    Purpose: Verify define_reference_list_for_internal uses title when available.
-    Why this matters: Title provides meaningful name for internal document references.
-    Setup summary: Create internal content chunk with title, call define_reference_list_for_internal, verify name is title.
-    """
-    # Arrange
-    content_chunks = [internal_content_chunk_with_title]
-    data = []
-
-    # Act
-    references = MessageStepLogger.define_reference_list_for_internal(
-        source="internal", content_chunks=content_chunks, data=data
-    )
-
-    # Assert
-    assert isinstance(references[0].name, str)
-    assert references[0].name == "Internal Document Title"
-
-
-@pytest.mark.ai
-def test_define_reference_list_for_internal__sets_name__from_chunk_key_when_no_title_AI(
-    internal_content_chunk_without_title: ContentChunk,
-) -> None:
-    """
-    Purpose: Verify define_reference_list_for_internal falls back to key when title missing.
-    Why this matters: References need names even when title is unavailable.
-    Setup summary: Create internal content chunk without title, call define_reference_list_for_internal, verify name is key.
-    """
-    # Arrange
-    content_chunks = [internal_content_chunk_without_title]
-    data = []
-
-    # Act
-    references = MessageStepLogger.define_reference_list_for_internal(
-        source="internal", content_chunks=content_chunks, data=data
-    )
-
-    # Assert
-    assert isinstance(references[0].name, str)
-    assert references[0].name == "doc_key_2"
-
-
-@pytest.mark.ai
-def test_define_reference_list_for_internal__sets_empty_url__for_internal_chunks_AI(
-    internal_content_chunk_with_title: ContentChunk,
-) -> None:
-    """
-    Purpose: Verify define_reference_list_for_internal sets empty URL for internal chunks.
-    Why this matters: Internal documents don't have URLs, only source IDs.
-    Setup summary: Create internal content chunk, call define_reference_list_for_internal, verify URL is empty string.
-    """
-    # Arrange
-    content_chunks = [internal_content_chunk_with_title]
-    data = []
-
-    # Act
-    references = MessageStepLogger.define_reference_list_for_internal(
-        source="internal", content_chunks=content_chunks, data=data
-    )
-
-    # Assert
-    assert isinstance(references[0].url, str)
-    assert references[0].url == ""
 
 
 # Full Message Tests - Web Search
